if isSpringSimpleProperty is detected (implicit RequestParam) set required=true by default.
Added logic to ensure required evaluates to true if at least one source defines it as true.
