+++ openapi-maven-plugin/src/test/java/io/github/kbuntrock/SpringClassAnalyserTest.java
@@ -2,7 +2,11 @@ package io.github.kbuntrock;
 
 import io.github.classgraph.ClassGraph;
 import io.github.classgraph.ScanResult;
-import io.github.kbuntrock.configuration.*;
+import io.github.kbuntrock.configuration.ApiConfiguration;
+import io.github.kbuntrock.configuration.CommonApiConfiguration;
+import io.github.kbuntrock.configuration.JavadocConfiguration;
+import io.github.kbuntrock.configuration.OperationIdHelper;
+import io.github.kbuntrock.configuration.Substitution;
 import io.github.kbuntrock.configuration.library.TagAnnotation;
 import io.github.kbuntrock.context.ApiContext;
 import io.github.kbuntrock.context.ProjectContext;
@@ -13,11 +17,38 @@ import io.github.kbuntrock.resources.endpoint.annotation.AnnotatedController;
 import io.github.kbuntrock.resources.endpoint.collection.CollectionController;
 import io.github.kbuntrock.resources.endpoint.collision.FirstEndpoint;
 import io.github.kbuntrock.resources.endpoint.collision.SecondEndpoint;
-import io.github.kbuntrock.resources.endpoint.enumeration.*;
+import io.github.kbuntrock.resources.endpoint.enumeration.TestEnumeration1Controller;
+import io.github.kbuntrock.resources.endpoint.enumeration.TestEnumeration2Controller;
+import io.github.kbuntrock.resources.endpoint.enumeration.TestEnumeration3Controller;
+import io.github.kbuntrock.resources.endpoint.enumeration.TestEnumeration4Controller;
+import io.github.kbuntrock.resources.endpoint.enumeration.TestEnumeration5Controller;
+import io.github.kbuntrock.resources.endpoint.enumeration.TestEnumeration6Controller;
 import io.github.kbuntrock.resources.endpoint.error.SameOperationController;
 import io.github.kbuntrock.resources.endpoint.file.FileUploadController;
 import io.github.kbuntrock.resources.endpoint.file.StreamResponseController;
-import io.github.kbuntrock.resources.endpoint.generic.*;
+import io.github.kbuntrock.resources.endpoint.generic.ExtendsGenericObjectMap;
+import io.github.kbuntrock.resources.endpoint.generic.ExtendsList;
+import io.github.kbuntrock.resources.endpoint.generic.ExtendsListV2;
+import io.github.kbuntrock.resources.endpoint.generic.ExtendsMap;
+import io.github.kbuntrock.resources.endpoint.generic.GenericDataController;
+import io.github.kbuntrock.resources.endpoint.generic.GenericMappingObject;
+import io.github.kbuntrock.resources.endpoint.generic.GenericityTestEight;
+import io.github.kbuntrock.resources.endpoint.generic.GenericityTestEleven;
+import io.github.kbuntrock.resources.endpoint.generic.GenericityTestFive;
+import io.github.kbuntrock.resources.endpoint.generic.GenericityTestFour;
+import io.github.kbuntrock.resources.endpoint.generic.GenericityTestNine;
+import io.github.kbuntrock.resources.endpoint.generic.GenericityTestOne;
+import io.github.kbuntrock.resources.endpoint.generic.GenericityTestSeven;
+import io.github.kbuntrock.resources.endpoint.generic.GenericityTestSix;
+import io.github.kbuntrock.resources.endpoint.generic.GenericityTestTen;
+import io.github.kbuntrock.resources.endpoint.generic.GenericityTestThree;
+import io.github.kbuntrock.resources.endpoint.generic.GenericityTestTwelve;
+import io.github.kbuntrock.resources.endpoint.generic.GenericityTestTwo;
+import io.github.kbuntrock.resources.endpoint.generic.Issue144;
+import io.github.kbuntrock.resources.endpoint.generic.Issue144ByInterface;
+import io.github.kbuntrock.resources.endpoint.generic.Issue89;
+import io.github.kbuntrock.resources.endpoint.generic.Issue95;
+import io.github.kbuntrock.resources.endpoint.generic.MappingObject;
 import io.github.kbuntrock.resources.endpoint.header.MultipartFileWithHeaderController;
 import io.github.kbuntrock.resources.endpoint.ignore.JsonIgnoreController;
 import io.github.kbuntrock.resources.endpoint.interfacedto.InterfaceController;
@@ -25,12 +56,6 @@ import io.github.kbuntrock.resources.endpoint.issues.Issue138;
 import io.github.kbuntrock.resources.endpoint.issues.Issue246;
 import io.github.kbuntrock.resources.endpoint.issues.Issue247;
 import io.github.kbuntrock.resources.endpoint.issues.Issue262;
-import io.github.kbuntrock.resources.endpoint.issues.issue332.OneParametersOutTwoParametersInNoMerge;
-import io.github.kbuntrock.resources.endpoint.issues.issue332.TwoParametersOutTwoParametersEmptyInWithAllMerged;
-import io.github.kbuntrock.resources.endpoint.issues.issue332.TwoParametersOutTwoParametersEmptyInWithPriority;
-import io.github.kbuntrock.resources.endpoint.issues.issue332.TwoParametersOutTwoParametersInNoMerge;
-import io.github.kbuntrock.resources.endpoint.issues.issue332.TwoParametersOutTwoParametersInWithAllMerged;
-import io.github.kbuntrock.resources.endpoint.issues.issue332.TwoParametersOutTwoParametersInWithMerge;
 import io.github.kbuntrock.resources.endpoint.jackson.JacksonJsonPropertyController;
 import io.github.kbuntrock.resources.endpoint.map.MapController;
 import io.github.kbuntrock.resources.endpoint.multipartformdata.MultipartFormDataController;
@@ -47,7 +72,12 @@ import io.github.kbuntrock.resources.endpoint.queryparam.EmptyValueParameterCont
 import io.github.kbuntrock.resources.endpoint.queryparam.QueryParamDtoBindingController;
 import io.github.kbuntrock.resources.endpoint.queryparam.QueryParamFlatMixNestedDtoBindingController;
 import io.github.kbuntrock.resources.endpoint.queryparam.QueryParamInMappingController;
-import io.github.kbuntrock.resources.endpoint.recursive.*;
+import io.github.kbuntrock.resources.endpoint.recursive.GenericRecursiveDtoController;
+import io.github.kbuntrock.resources.endpoint.recursive.GenericRecursiveInterfaceDtoController;
+import io.github.kbuntrock.resources.endpoint.recursive.GenericRecursiveInterfaceListDtoInParameterController;
+import io.github.kbuntrock.resources.endpoint.recursive.GenericRecursiveListDtoController;
+import io.github.kbuntrock.resources.endpoint.recursive.RecursiveDtoController;
+import io.github.kbuntrock.resources.endpoint.recursive.RecursiveDtoInParameterController;
 import io.github.kbuntrock.resources.endpoint.spring.OptionalController;
 import io.github.kbuntrock.resources.endpoint.spring.ResponseEntityController;
 import io.github.kbuntrock.resources.endpoint.spring.ResponseEntityUnparametrizedController;
@@ -70,7 +100,12 @@ import java.io.FileInputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.nio.file.Files;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Optional;
 
 public class SpringClassAnalyserTest extends AbstractTest {
 
@@ -1016,45 +1051,6 @@ public class SpringClassAnalyserTest extends AbstractTest {
 	@Test
 	public void empty_value_query_parameter() throws MojoFailureException, IOException, MojoExecutionException {
 		final DocumentationMojo mojo = createBasicMojo(EmptyValueParameterController.class.getCanonicalName());
-		checkGenerationResult(mojo.documentProject());
-	}
-
-	@Test
-	public void OneParametersOutTwoParametersInNoMerge() throws MojoFailureException, IOException, MojoExecutionException {
-		final DocumentationMojo mojo = createBasicMojo(OneParametersOutTwoParametersInNoMerge.class.getCanonicalName());
-		checkGenerationResult(mojo.documentProject());
-	}
-
-	@Test
-	public void TwoParametersOutTwoParametersEmptyInWithAllMerged()
-		throws MojoFailureException, IOException, MojoExecutionException {
-		final DocumentationMojo mojo = createBasicMojo(
-			TwoParametersOutTwoParametersEmptyInWithAllMerged.class.getCanonicalName());
-		checkGenerationResult(mojo.documentProject());
-	}
-
-	@Test
-	public void TwoParametersOutTwoParametersInNoMerge() throws MojoFailureException, IOException, MojoExecutionException {
-		final DocumentationMojo mojo = createBasicMojo(TwoParametersOutTwoParametersInNoMerge.class.getCanonicalName());
-		checkGenerationResult(mojo.documentProject());
-	}
-
-	@Test
-	public void TwoParametersOutTwoParametersInWithAllMerged() throws MojoFailureException, IOException, MojoExecutionException {
-		final DocumentationMojo mojo = createBasicMojo(TwoParametersOutTwoParametersInWithAllMerged.class.getCanonicalName());
-		checkGenerationResult(mojo.documentProject());
-	}
-
-	@Test
-	public void TwoParametersOutTwoParametersInWithMerge() throws MojoFailureException, IOException, MojoExecutionException {
-		final DocumentationMojo mojo = createBasicMojo(TwoParametersOutTwoParametersInWithMerge.class.getCanonicalName());
-		checkGenerationResult(mojo.documentProject());
-	}
-
-	@Test
-	public void TwoParametersOutTwoParametersEmptyInWithPriority()
-		throws MojoFailureException, IOException, MojoExecutionException {
-		final DocumentationMojo mojo = createBasicMojo(TwoParametersOutTwoParametersEmptyInWithPriority.class.getCanonicalName());
 		checkGenerationResult(mojo.documentProject());
 	}
