+++ openapi-maven-plugin/src/test/java/io/github/kbuntrock/resources/endpoint/issues/issue332/TwoParametersOutTwoParametersEmptyInWithAllMerged.java
@@ -2,9 +2,7 @@ package io.github.kbuntrock.resources.endpoint.issues.issue332;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
-import io.swagger.v3.oas.annotations.enums.Explode;
 import io.swagger.v3.oas.annotations.enums.ParameterIn;
-import io.swagger.v3.oas.annotations.enums.ParameterStyle;
 import io.swagger.v3.oas.annotations.media.ArraySchema;
 import io.swagger.v3.oas.annotations.media.Schema;
 import org.springframework.http.MediaType;
